Move stuff to sub folder, update with small refactor
