+++ packages/core/src/utils/openai/constants.ts
